Note 2a for simple MCMC simulations
The note provides technical information and explanations on the use of BayesVL in performing MCMC simulations for simpler models (non-hierarchical). It employs a real data set surveying 1042 patients about their financial conditions, use of debt for financing healthcare treatments. It will be part of the guide book for the BayesVL R package